Fix input fields after toggle
If you toggled the manual address entry, the combo fields will show both
`select` and `input` fields. This patch changes the behavior, so that
the combo fields will be set to a sensible value on every toggle.

Sensible value means that `input` fields that have not any text in them
are hidden and replaced by the `select` fields.